lesson3/workspace/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# Observability
langsmith>=0.2.0

# Web Search
tavily-python>=0.5.0

# HTTP client
httpx>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Type checking (optional but recommended)
mypy>=1.8.0
